Refactored Html parser - part 1
