+++ src/Razor/src/Microsoft.AspNetCore.Razor.Language/Legacy/ParserContext.cs
@@ -49,6 +49,8 @@ namespace Microsoft.AspNetCore.Razor.Language.Legacy
 
         public bool InTemplateContext { get; set; }
 
+        public bool StartOfLine { get; set; }
+
         public AcceptedCharactersInternal LastAcceptedCharacters { get; set; } = AcceptedCharactersInternal.None;
 
         public bool EndOfFile
